sth changed

+++ kinect.cpp
@@ -11,7 +11,7 @@ static Mat create_depth_vis(Mat depth){
 		for (int i = 0; i < depth.rows;i++)
 		for (int j = 0; j < depth.cols; j++)
 		{
-			uint16_t val = depth.at<uint16_t>(i,j);
+			int val = depth.at<int>(i,j);
 			char gray_val = 255-val >> 8;
 			gray.at<char>(i, j) = gray_val;
 		}
